test(publisher): split chain event lane suites by responsibility
