+++ packages/publisher/test/helpers/chain-event-lane-fixture.ts
@@ -1,0 +1,53 @@
+import { OxigraphStore } from '@origintrail-official/dkg-storage';
+import { TypedEventBus } from '@origintrail-official/dkg-core';
+import type { ChainAdapter, ChainEvent, EventFilter } from '@origintrail-official/dkg-chain';
+import { PublishHandler } from '../../src/publish-handler.js';
+import type { JournalEntry } from '../../src/publish-journal.js';
+
+export function makeChain(head: number, events: ChainEvent[]): {
+  adapter: ChainAdapter;
+  filters: EventFilter[];
+} {
+  const filters: EventFilter[] = [];
+  const adapter = {
+    chainId: 'mock:0',
+    getBlockNumber: async () => head,
+    listenForEvents: async function* (f: EventFilter): AsyncIterable<ChainEvent> {
+      filters.push(f);
+      const fromBlock = f.fromBlock ?? 0;
+      const toBlock = f.toBlock ?? Number.MAX_SAFE_INTEGER;
+      for (const evt of events) {
+        if (f.eventTypes.includes(evt.type) && evt.blockNumber >= fromBlock && evt.blockNumber <= toBlock) {
+          yield evt;
+        }
+      }
+    },
+  } as unknown as ChainAdapter;
+  return { adapter, filters };
+}
+
+export function makeHandler(): PublishHandler {
+  return new PublishHandler(new OxigraphStore(), new TypedEventBus());
+}
+
+export function markPending(handler: PublishHandler, restoredFromJournal: boolean): void {
+  (handler as unknown as { pendingPublishes: Map<string, unknown> }).pendingPublishes.set(
+    restoredFromJournal ? 'restored' : 'live',
+    { expectedMerkleRoot: new Uint8Array(32), restoredFromJournal },
+  );
+}
+
+export function journalEntry(overrides: Partial<JournalEntry> = {}): JournalEntry {
+  return {
+    ual: 'did:dkg:mock:0/0xa1a1a1a1a1a1a1a1a1a1a1a1a1a1a1a1a1a1a1a1/1',
+    contextGraphId: 'contextGraph-1',
+    expectedPublisherAddress: '0x' + 'a1'.repeat(20),
+    expectedMerkleRoot: '0x' + '55'.repeat(32),
+    expectedStartKAId: '1',
+    expectedEndKAId: '1',
+    expectedChainId: 'mock:0',
+    rootEntities: [],
+    createdAt: Date.now(),
+    ...overrides,
+  };
+}
